+++ ext/intl/intl_data.h
@@ -24,8 +24,8 @@
 #include "intl_error.h"
 
 /* Mock object to generalize error handling in sub-modules.
-   Sub-module data structures should always have error as first element 
-   for this to work! 
+   Sub-module data structures should always have error as first element
+   for this to work!
 */
 typedef struct _intl_data {
 	intl_error		error;
@@ -35,7 +35,7 @@ typedef struct _intl_data {
 #define INTL_METHOD_INIT_VARS(oclass, obj)		\
 	zval*             object  = NULL;			\
 	oclass##_object*  obj     = NULL;			\
-	intl_error_reset( NULL );			
+	intl_error_reset( NULL );
 
 #define INTL_DATA_ERROR(obj)				(((intl_object *)(obj))->error)
 #define INTL_DATA_ERROR_P(obj)				(&(INTL_DATA_ERROR((obj))))
@@ -81,7 +81,7 @@ typedef struct _intl_data {
 #define INTL_METHOD_RETVAL_UTF8(obj, ustring, ulen, free_it)									\
 {																								\
 	char *u8value;																				\
-	int u8len;																					\
+	size_t u8len;																				\
 	intl_convert_utf16_to_utf8(&u8value, &u8len, ustring, ulen, &INTL_DATA_ERROR_CODE((obj)));	\
 	if((free_it)) {																				\
 		efree(ustring);																			\
@@ -96,14 +96,14 @@ typedef struct _intl_data {
 #define INTL_CHECK_LOCALE_LEN(locale_len)												\
 	if((locale_len) > INTL_MAX_LOCALE_LEN) {											\
 		intl_error_set( NULL, U_ILLEGAL_ARGUMENT_ERROR,									\
-	"Locale string too long, should be no longer than 80 characters", 0 );	\
+		"Locale string too long, should be no longer than 80 characters", 0 );			\
 		RETURN_NULL();																	\
 	}
 
 #define INTL_CHECK_LOCALE_LEN_OBJ(locale_len, object)									\
 	if((locale_len) > INTL_MAX_LOCALE_LEN) {											\
 		intl_error_set( NULL, U_ILLEGAL_ARGUMENT_ERROR,									\
-	"Locale string too long, should be no longer than 80 characters", 0 );	\
+		"Locale string too long, should be no longer than 80 characters", 0 );			\
 		zval_dtor(object);																\
 		ZVAL_NULL(object);																\
 		RETURN_NULL();																	\
